+++ ax3l/constants/DReportMgr.py
@@ -2,6 +2,7 @@ from typing import Final
 
 
 class DReportMgr:
+    SIMULATION_BUCKET_SIZE: Final[int] = 20
     REQUEST_TIMEOUT_SECONDS: Final[int] = 30
     PORT: Final[int] = 28870
     PORT_DEV: Final[int] = 28868